refactor: use model in cart repository in service
