+++ src/api/profile/cart/cart.repository.ts
@@ -1,77 +1,53 @@
 import crypto from "node:crypto";
 
-import DB from "../../../db";
+import Cart from "./cart.model";
 
-import type { Cart, CartItem, UpdateCartItem } from "./cart.types";
+import type { UpdateCartItem } from "./cart.types";
 import type { UserId } from "../profile.type";
 import type { Product } from "../../products/product/product.types";
 
-export const getCart = async (userId: UserId): Promise<Cart> => {
-  return new Promise((res, _) => {
-    let cart: Cart = DB.get("carts").find(
-      (cart: Cart) => cart.userId === userId && !cart.isDeleted
-    );
-    if (!cart) {
-      cart = {
-        id: crypto.randomUUID(),
-        userId,
-        isDeleted: false,
-        items: [],
-      };
-      DB.get("carts").push(cart);
-    }
-    return res(cart);
-  });
-};
+export const getCart = async (userId: UserId) => {
+  const cart = await Cart.findOne({ userId, isDeleted: false });
+  if (!cart) {
+    const newCart = new Cart({
+      id: crypto.randomUUID(),
+      userId,
+      isDeleted: false,
+      items: [],
+    });
+    await newCart.save();
+    return newCart;
+  }
+  return cart;
+}
+
 
 export const updateCart = async (
   userId: UserId,
   updateCartData: UpdateCartItem[],
   products: Product[] | []
-): Promise<Cart | null> => {
-  return new Promise(async (res, _) => {
-    const carts: Cart[] = DB.get("carts");
-    const cartIndex = carts.findIndex(
-      (cart: Cart) => cart.userId === userId && !cart.isDeleted
-    );
-    if (cartIndex === -1) {
-      return res(null);
-    }
-
-    const mappedProducts: CartItem[] = products.map(
-      (product: Product, index) => {
-        return {
-          product,
-          count: updateCartData[index].count,
-        };
-      }
-    );
-
-    carts[cartIndex] = {
-      ...carts[cartIndex],
-      items: [...mappedProducts],
+) => {
+  const cart = await Cart.findOne({ userId, isDeleted: false });
+  if (!cart) {
+    return;
+  }
+  const mappedProducts = products.map((product, index) => {
+    return {
+      product,
+      count: updateCartData[index].count,
     };
-
-    return res({ ...carts[cartIndex] });
   });
-};
-
-export const deleteCart = async (userId: UserId): Promise<Cart | null> => {
-  return new Promise((res, _) => {
-    const carts: Cart[] = DB.get("carts");
-    const cartIndex = carts.findIndex(
-      (cart: Cart) => cart.userId === userId && !cart.isDeleted
-    );
-
-    if (cartIndex === -1) {
-      res(null);
-    }
-
-    carts[cartIndex] = {
-      ...carts[cartIndex],
-      items: [],
-      isDeleted: true,
-    };
-    res({ ...carts[cartIndex] });
-  });
+  cart.items = mappedProducts;
+  await cart.save();
+  return cart;
+}
+  
+export const deleteCart = async (userId: UserId) => {
+  const cart = await Cart.findOne({ userId, isDeleted: false });
+  if (!cart) {
+    return;
+  }
+  cart.isDeleted = true;
+  await cart.save();
+  return cart;
 };
